requirements.txt
fastapi
uvicorn
openai
ffmpeg
PyMuPDF
PyPDF2  
python-multipart
pymongopy